IW-NeRF code
We have designed a watermarking framework that uses Implicit Neural Representation (INR) for copyright protection. After implicitly representing the watermark information, we embed the network representing the watermark information into the network of the NeRF model by designing a specific key. The copyright validator extracts watermarks from the NeRF model using a key to prove the copyright of the NeRF model.